LLMs/2-Assistant_QA/requirements.txt
python_doctr==0.8.1
PyMuPDF==1.24.9
langchain_chroma==0.1.2
llama_index==0.10.59
llama_index_embeddings_huggingface==0.2.2
langchain_community==0.2.11
typing-inspect==0.8.0
typing_extensions==4.8.0
pydantic==1.10.11
langchain_huggingface==0.0.3
huggingface-hub==0.24.5
pysqlite3_binary==0.5.3
chromadb==0.4.4
tf2onnx==1.16.1
streamlit==1.37.0
streamlit-pdf-viewer==0.0.16
#requests==2.28.2